refactor(utils): promote gamelog_parser to services/gamelog_service (#444)
Split the 728-line utils/gamelog_parser.py into services/gamelog_service/
across usernames, discovery, parser, formats, and service modules. Per
the issue, this is a service-shaped responsibility (parsing/extraction/
file discovery/format detection), not a generic util.

Updates the 2 callers (widgets/frames/match_history/handlers.py and
tests/test_gamelog_parser.py). utils/gamelog_parser.py deleted.
